refactor(metrics): move isPublicEndpointUrl to background API
