+++ ui/hooks/useNetworkConnectionBanner.test.ts
@@ -65,6 +65,13 @@ jest.mock('../../shared/modules/selectors/networks', () => {
   };
 });
 
+jest.mock('../store/background-connection', () => {
+  return {
+    ...jest.requireActual('../store/background-connection'),
+    submitRequestToBackground: jest.fn().mockResolvedValue(true),
+  };
+});
+
 const mockSelectFirstUnavailableEvmNetwork = jest.mocked(
   selectFirstUnavailableEvmNetwork,
 );
@@ -186,7 +193,7 @@ describe('useNetworkConnectionBanner', () => {
           });
         });
 
-        it('creates a MetaMetrics event to capture that the status changed', () => {
+        it('creates a MetaMetrics event to capture that the status changed', async () => {
           mockSelectFirstUnavailableEvmNetwork.mockReturnValue({
             networkName: 'Ethereum Mainnet',
             networkClientId: 'mainnet',
@@ -204,8 +211,10 @@ describe('useNetworkConnectionBanner', () => {
             undefined,
             () => mockTrackEvent,
           );
-          act(() => {
+          await act(async () => {
             jest.advanceTimersByTime(5000);
+            // Flush microtask queue to allow async trackNetworkBannerEvent to complete
+            await Promise.resolve();
           });
 
           expect(mockTrackEvent).toHaveBeenCalledWith({
@@ -293,7 +302,7 @@ describe('useNetworkConnectionBanner', () => {
         });
       });
 
-      it('creates a MetaMetrics event to capture that the status changed', () => {
+      it('creates a MetaMetrics event to capture that the status changed', async () => {
         mockSelectFirstUnavailableEvmNetwork.mockReturnValue({
           networkName: 'Ethereum Mainnet',
           networkClientId: 'mainnet',
@@ -318,8 +327,10 @@ describe('useNetworkConnectionBanner', () => {
           undefined,
           () => mockTrackEvent,
         );
-        act(() => {
+        await act(async () => {
           jest.advanceTimersByTime(25000);
+          // Flush microtask queue to allow async trackNetworkBannerEvent to complete
+          await Promise.resolve();
         });
 
         expect(mockTrackEvent).toHaveBeenCalledWith({
